+++ server/src/groups/groups.controller.ts
@@ -1,35 +1,58 @@
-import { Controller, Get, Post,Patch, Delete, Param, Body } from '@nestjs/common';
+import { Controller, Get, Post,Patch, Delete, Param, Body, Request, UseGuards, HttpException } from '@nestjs/common';
 import { GroupsService } from './groups.service';
-import { Prisma } from '@prisma/client';
 import { CreateGroupDto } from './dto/create-groupe.dto';
 import { UpdateGroupDto } from './dto/update-groupe.dto';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
 
+@UseGuards(JwtAuthGuard) 
 @Controller('groups')
 export class GroupsController {
   constructor(private readonly groupsService: GroupsService) {}
 
   @Get()
-  async getGroups(){
-    return await this.groupsService.getGroups()
+  async getGroups(@Request() req){
+    return await this.groupsService.getGroups(req.user.id)
   }
 
   @Get("/:id")
-  async getGroup(@Param("id") id: number){
-    return await this.groupsService.getGroup(id)
+  async getGroup(@Param("id") id: number, @Request() req){
+    const userGroup = await this.groupsService.getGroup(id);
+
+    if(req.user.id === userGroup.user_id){
+        return userGroup;
+    }  
+    else{
+        throw new HttpException('Unauthorized', 401);
+    }
   }
 
   @Post()
-  async createGroup(@Body() groupDto: CreateGroupDto){
-    return await this.groupsService.createGroup(groupDto)
+  async createGroup(@Body() groupDto: CreateGroupDto, @Request() req){
+    if(req.user.id === groupDto.user_id){
+      return await this.groupsService.createGroup(groupDto)
+    }
+    else {
+      throw new HttpException('Unauthorized', 401);
+    }
   }
 
   @Patch("/:id")
-  async updateGroup(@Param("id") id: number, @Body() groupDto: UpdateGroupDto){
-    return await this.groupsService.updateGroup(id, groupDto)
+  async updateGroup(@Param("id") id: number, @Body() groupDto: UpdateGroupDto, @Request() req){
+    const userGroup = await this.groupsService.getGroup(id);
+
+    if(req.user.id !== userGroup.user_id){
+      throw new HttpException('Unauthorized', 401);
+    }
+    return await this.groupsService.updateGroup(id, groupDto);
   }
 
   @Delete("/:id")
-  async deleteGroup(@Param("id") id: number){
-    return await this.groupsService.deleteGroup(id)
+  async deleteGroup(@Param("id") id: number, @Request() req){
+    const userGroup = await this.groupsService.getGroup(id);
+
+    if(req.user.id !== userGroup.user_id){
+      throw new HttpException('Unauthorized', 401);
+    }    
+    return await this.groupsService.deleteGroup(id);
   }
 }
